requirements.txt
Django==1.10.3
psycopg2==2.6.2
serpy==0.1.1
djangorestframework==3.5.3
markdown==2.6.7
django-filter==1.0.0